+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added 'Serialize/Deserialize' trait bounds to `Action<C>`
+
 ## [0.20.1] - 2025-12-29
 
 ### Changed